requirements.txt
Flask==2.0.3
tensorflow==2.10.0
gunicorn==20.1.0
numpy==1.21.4
pillow==8.4.0
